// dashboard/src/utils/hasura-api/generated/schemas/deleteRemoteRelationshipBulkOperation.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { DeleteRemoteRelationshipBulkOperationType } from './deleteRemoteRelationshipBulkOperationType';
import type { DeleteRemoteRelationshipStep } from './deleteRemoteRelationshipStep';

export interface DeleteRemoteRelationshipBulkOperation {
  /** Type of operation to delete remote relationship */
  type: DeleteRemoteRelationshipBulkOperationType;
  /** Resource version of the operation */
  resource_version: number;
  args: DeleteRemoteRelationshipStep[];
}