// dashboard/src/utils/hasura-api/generated/schemas/dropRelationshipBulkOperation.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { DropRelationshipBulkOperationType } from './dropRelationshipBulkOperationType';
import type { DropRelationshipStep } from './dropRelationshipStep';

export interface DropRelationshipBulkOperation {
  /** Type of operation to drop relationship */
  type: DropRelationshipBulkOperationType;
  /** Resource version of the operation */
  resource_version: number;
  args: DropRelationshipStep[];
}